requirements.txt
asgiref==3.11.0
crispy-bootstrap4==2025.6
Django==5.2.9
django-crispy-forms==2.5
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
tzdata==2025.2
